backend/requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
mysqlclient
python-dotenvpython